refactor(proxy): move DDSpanTagger to its own file litellm/proxy/dd_span_tagger.py
